Example refactoring to share counters logic
This was my experiment to demonstrate how we can decompose our collection logic into pipelines and share it across multiple 'agents'. In this case think of dotnet-counters and dotnet-monitor are two different implementations of an agent. dotnet-counters only knows how to do counter collection and has console UI whereas dotnet-monitor knows how to do many kinds of collection and has REST api. Now both agents are sharing the collection logic to handle reading counters via EventPipe, parsing it, and emitting it as either csv or json format streams. I used IPipeline as the basic unit collection behavior. The interface is deliberately pretty abstract representing it only as something that can be started and stopped while exposing no detail about what data is being collected, how it flows, or where it is going. I'm not too attached to the exact API choices though. In this case I put the implementation of EventPipeCounterPipeline into Microsoft.Diagnostics.Monitoring but there is no requirement that all pipeline implementations be in that assembly vs. a separate one.

There are a few ways we could build on top of this. I don't think any of the ideas below are requirements for an initial MVP, more like suggestions for a longer roadmap.

1. We could imagine repeating this pattern with other of the dotnet-* diagnostic tools, or perhaps more importantly use similar patterns to share more advanced collectors between dotnet-monitor and dedicated agents that Microsoft wants to deploy into Azure environments. If the collectors are in different assemblies it should be easy to pick and choose the set of collection pipelines we want each agent to have available.

2. The counter pipeline is probably not too hard to generalize/subclass to handle all the EventPipe cases. Probably the trickiest aspects are getting the exception handling and shutdown mechanics correct (I expect the current code still has some bugs there). Those elements should be the same regardless of what data is being pulled through EventPipe so it would be nice to share it.

3. Right now creating a pipeline requires the agent to use a strongly typed managed API to invoke the constructor and pass initial configuration. However if we created an IPipelineFactory we could instantiate pipelines using loosely typed configuration information such as an instance of IConfiguration or a snippet of serialized text. This separates the config parsing from the REST api and lets us do a few things:
a) Have general purpose REST apis that could setup any pipeline that has a factory
b) Test config parsing independent of the REST api
c) Make agents that are configured with an adjacent config file instead of a REST api
d) Describe more complex multi-pipeline configurations, such as embedding the description of a trigger pipeline inside a triggerable pipeline.

4. We could make the agent more explicitly track the set of pipelines that are currently running to provide agent self-diagnostics and management. For example a pipeline could be explicitly started, stopped, or modified while it was running via REST apis. We could also expose pipeline warnings, error messages, resource quotas, and counters for resource usage, throughput and latency.